Try a bunch of different things to share with others
Here is where I am stuck one might hope that the `sleep 10` call at
game.rb:168 would not block the main thread, however, it does...  This
indicates to me the the concurrent lib does not use the classic ruby
threads.

The desired outcome is to come up with the code that will allow the code
at game.rb:169 to work without needing to pass a block into the
`select_a_card` method. Essentially remove the continuation so that the
code is a bit more readable. It has to do this without blocking the main
thread or else the gosu framework will not continue to draw the GUI.