Split play_card and post_card_play_clean_up
The main reason to do this is in the lifecycle of the game we should
check if there is a winner before enforcing limits, shuffling the deck,
replenishing hand and checking if the turn is over. This can't all be
done together since each interface (TUI/GUI) will want to handle it
differently.